hot fixes: -UserRestController.getUser(?) has changed authorization`s logic; -WorkoutService got duration min value checker - org.api.workout.controller.dto and *.exceptions were refactored to org.api.workout.dto and *.exceptions
