+++ src/main/java/org/api/workout/controllers/error/CustomErrorController.java
@@ -1,0 +1,29 @@
+package org.api.workout.controllers.error;
+
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.boot.web.servlet.error.ErrorController;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Component
+public class CustomErrorController implements ErrorController {
+
+    @RequestMapping("/error")
+    @ResponseBody
+    public ResponseEntity<Map<String, Object>> handleError(HttpServletRequest request) {
+        Object status = request.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
+
+        Map<String, Object> body = new HashMap<>();
+        body.put("error", "Something went wrong");
+        body.put("status", status);
+
+        return new ResponseEntity<>(body, HttpStatus.valueOf((int) status));
+    }
+}
