+++ src/main/java/org/api/workout/dto/user/UserDTO.java
@@ -1,8 +1,8 @@
-package org.api.workout.controllers.dto.user;
+package org.api.workout.dto.user;
 
 
-import org.api.workout.controllers.dto.goal.GoalDTO;
-import org.api.workout.controllers.dto.workout.WorkoutDTO;
+import org.api.workout.dto.goal.GoalDTO;
+import org.api.workout.dto.workout.WorkoutDTO;
 
 import java.time.LocalDateTime;
 import java.util.List;
