+++ src/main/java/org/api/workout/exceptions/goals/GoalNotFoundException.java
@@ -1,4 +1,4 @@
-package org.api.workout.controllers.exceptions.goals;
+package org.api.workout.exceptions.goals;
 
 public class GoalNotFoundException extends RuntimeException {
     public GoalNotFoundException(String message) {
