+++ src/main/java/org/api/workout/exceptions/user/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package org.api.workout.controllers.exceptions.user;
+package org.api.workout.exceptions.user;
 
 public class UserAlreadyExistsException extends RuntimeException {
     public UserAlreadyExistsException(String message) {
