+++ src/main/java/org/api/workout/exceptions/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package org.api.workout.controllers.exceptions.user;
+package org.api.workout.exceptions.user;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
