+++ src/main/java/org/api/workout/exceptions/workout/AccessForbiddenException.java
@@ -1,4 +1,4 @@
-package org.api.workout.controllers.exceptions.workout;
+package org.api.workout.exceptions.workout;
 
 public class AccessForbiddenException extends RuntimeException {
     public AccessForbiddenException(String message) {
