+++ src/main/resources/application.properties
@@ -10,3 +10,5 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 
+
+
